refactor(topology): typed proximity, bin and neighborhood depth
Replaces bare-u8 PO/bin math with `ProximityOrder` and `Bin` newtypes
and introduces `NeighborhoodDepth` with `recompute` matching bee's
`kademlia.go:896-920` recalcDepth algorithm.

- `SwarmAddressExt` gains `proximity_order`, `xor`, `bin`.
- `ProximityIndex` keyed by `Bin` not `u8`.
- `NeighborhoodDepth(Bin)` with `recompute(connected_per_bin, saturation, low_watermark)`.
- Golden test vectors from bee `proximity_test.go`.

Unit 13 of the bee-mainnet bootnode interop plan.

+++ crates/swarm/bandwidth/pricing/src/fixed.rs
@@ -39,7 +39,7 @@ impl<S: SwarmSpec> Pricer for FixedPricer<S> {
     fn peer_price(&self, peer: &OverlayAddress, chunk: &ChunkAddress) -> u64 {
         let peer_addr: &SwarmAddress = peer;
         let chunk_addr: &SwarmAddress = chunk;
-        let proximity = peer_addr.proximity(chunk_addr);
+        let proximity = u8::from(peer_addr.proximity(chunk_addr));
         let factor = (self.spec.max_po() as u64) - (proximity as u64) + 1;
         factor * self.base_price
     }
